+++ kyuubi-hive-beeline/src/main/java/org/apache/hive/beeline/Commands.java
@@ -52,10 +52,10 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 import java.util.TreeSet;
-import org.apache.hadoop.hive.common.cli.ShellCmdExecutor;
 import org.apache.hadoop.io.IOUtils;
+import org.apache.hive.beeline.common.cli.ShellCmdExecutor;
+import org.apache.hive.beeline.common.util.HiveStringUtils;
 import org.apache.hive.beeline.logs.KyuubiBeelineInPlaceUpdateStream;
-import org.apache.hive.common.util.HiveStringUtils;
 import org.apache.kyuubi.jdbc.hive.JdbcConnectionParams;
 import org.apache.kyuubi.jdbc.hive.KyuubiStatement;
 import org.apache.kyuubi.jdbc.hive.Utils;
